Fixed XDF continuous polling bug, renamed tutorials
