+++ tutorials/CSP_tutorial.py
